keep-session: fixed bug on old version of mpv
on the older version of mpv my laptop had the idle-active property
would return "no" when initially booting up in idle mode
this disabled the auto-load feature of the script

Now, the script only checks to see if the player was booted with
a playlist of 0 length. If that is so it auto-loads the session.

Also in this release if you close mpv without any files in the playlist,
then it does not overide the old session