cycle-commands: changed command syntax again
I don't like breaking changes like this, but this should be the final
iteration of the syntax for this script.

The new syntax is significantly easier to write and understand because
it follows the exact same syntax as standard input.conf commands.

This fixes both the terrible comprehensability of the json syntax, and
the hackiness/non-robust parsing of my original syntax, which caused me
to switch to json for better maintainability.

This new syntax also allows me to use a different API call to run the
commands. This new function uses the same parser as the normal
input.conf commands, so things like property expansion, are
supported, and things will be written to the OSD by default.